Create NoteListItem component
